+++ src/senaite/core/content/instrumentlocations.py
@@ -18,6 +18,7 @@
 # Copyright 2018-2024 by it's authors.
 # Some rights reserved, see README and LICENSE.
 
+from bika.lims.interfaces import IDoNotSupportSnapshots
 from plone.dexterity.content import Container
 from plone.supermodel import model
 from senaite.core.interfaces import IHideActionsMenu
@@ -29,7 +30,7 @@ class IInstrumentLocations(model.Schema):
     """
 
 
-@implementer(IInstrumentLocations, IHideActionsMenu)
+@implementer(IInstrumentLocations, IDoNotSupportSnapshots, IHideActionsMenu)
 class InstrumentLocations(Container):
     """A container for instrument locations
     """