+++ CHANGES.rst
@@ -5,6 +5,11 @@ Changelog
 ------------------
 
 - #2476 Fix format type to include strings in calculation formulas
+- #2484 Disable CSRF protection in SENAITE
+- #2482 Added DurationField and DurationWidget for Dexterity types
+- #2481 Disable snapshots for container on content add
+- #2478 Migrate Sample Conditions to Dexterity
+- #2471 Migrate Lab Departments to Dexterit
 - #2469 Rename permission "Reject" to "senaite.core: Transition: Reject Analysis"
 - #2468 Make the prepublish guard to rely on interfaces instead of statuses
 - #2470 Fix min/max shifted on negative result and % error